events: add a parser for perf events
This supports the symbolic and PMU formats of perf, which are already
a horribly complicated syntax. It understands built-in events,
querying dynamic PMU events from /sys, and the symbolic CPU events
from "perf list -j" (if the perf command is very recent).

It doesn't support raw event, breakpoint, BPF, or group syntax.

There are several TODOs in the code, but it's quite functional.